refactor: update the strategies
